Reorganise new code for ungridded/ungridded colocation in different modules
